requirements.txt
Django==5.2.4
djangorestframework==3.16.0
python-slugify==8.0.4
mysqlclient==2.2.7
python-dotenv==1.1.1
